requirements.txt
langchain
openai
python-dotenv
faiss-cpu
wikipedia
wikipedia-api
langchain_community
langchain-openai

